move callback.py to bdpy.pipeline
