+++ bdpy/recon/torch/modules/critic.py
@@ -6,7 +6,7 @@ from typing import Dict, Iterable
 import torch
 import torch.nn as nn
 
-from bdpy.util.callback import CallbackHandler, BaseCallback
+from bdpy.pipeline.callback import CallbackHandler, BaseCallback
 
 
 _FeatureType = Dict[str, torch.Tensor]
